+++ tests/src/Unit/Resolver/WeeklyResolverTest.php
@@ -4,6 +4,8 @@ namespace Drupal\Tests\timesup\Unit\Resolver;
 
 use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Cache\CacheTagsInvalidatorInterface;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Config\ImmutableConfig;
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Logger\LoggerChannelInterface;
 use Drupal\Core\State\StateInterface;
@@ -41,6 +43,20 @@ class WeeklyResolverTest extends UnitTestCase {
   protected $state;
 
   /**
+   * The config factory service.
+   *
+   * @var \Drupal\Core\Config\ConfigFactoryInterface
+   */
+  protected $configFactory;
+
+  /**
+   * The mocked configuration object timesup.settings.
+   *
+   * @var \Drupal\Core\Config\ImmutableConfig
+   */
+  protected $settings;
+
+  /**
    * The logger.
    *
    * @var \Psr\Log\LoggerInterface
@@ -68,6 +84,11 @@ class WeeklyResolverTest extends UnitTestCase {
     parent::setUp();
     $this->cacheTagsInvalidator = $this->createMock(CacheTagsInvalidatorInterface::class);
 
+    // Create a mock config factory and config object.
+    $this->configFactory = $this->createMock(ConfigFactoryInterface::class);
+    $this->settings = $this->createMock(ImmutableConfig::class);
+    $this->configFactory->method('get')->with('timesup.settings')->willReturn($this->settings);
+
     $this->state = $this->createMock(StateInterface::class);
 
     $this->time = $this->createMock(TimeInterface::class);
@@ -83,7 +104,7 @@ class WeeklyResolverTest extends UnitTestCase {
    * @covers ::getCacheTags
    */
   public function testGetCacheTags() {
-    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->state, $this->time, $this->loggerFactory);
+    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->configFactory, $this->state, $this->time, $this->loggerFactory);
     $tags = $this->invokeMethod($resolver, 'getCacheTags');
     $this->assertSame(['timesup', 'timesup:weekly'], $tags);
   }
@@ -92,9 +113,25 @@ class WeeklyResolverTest extends UnitTestCase {
    * @covers ::getLastRunKey
    */
   public function testGetLastRunKey() {
-    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->state, $this->time, $this->loggerFactory);
+    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->configFactory, $this->state, $this->time, $this->loggerFactory);
     $key = $this->invokeMethod($resolver, 'getLastRunKey');
     $this->assertEquals('timesup.last_run.WeeklyResolver', $key);
+  }
+
+  /**
+   * @covers ::shouldApply
+   */
+  public function testShouldApplySettingsDisabled() {
+    $this->settings->expects($this->once())
+      ->method('get')->with('resolvers')->willReturn(['weekly' => FALSE]);
+
+    $this->state->expects($this->never())
+      ->method('get')->with('timesup.last_run.WeeklyResolver');
+    $this->time->expects($this->never())
+      ->method('getRequestTime');
+
+    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->configFactory, $this->state, $this->time, $this->loggerFactory);
+    $this->assertFalse($resolver->shouldApply());
   }
 
   /**
@@ -103,12 +140,15 @@ class WeeklyResolverTest extends UnitTestCase {
    * @dataProvider shouldApplyProvider
    */
   public function testShouldApply($request_time, $last_run, $expected) {
+    $this->settings->expects($this->once())
+      ->method('get')->with('resolvers')->willReturn(['weekly' => TRUE]);
+
     $this->state->expects($this->once())
       ->method('get')->with('timesup.last_run.WeeklyResolver')->willReturn($last_run);
     $this->time->expects($this->once())
       ->method('getRequestTime')->willReturn($request_time);
 
-    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->state, $this->time, $this->loggerFactory);
+    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->configFactory, $this->state, $this->time, $this->loggerFactory);
     $this->assertEquals($expected, $resolver->shouldApply());
   }
 
@@ -146,7 +186,7 @@ class WeeklyResolverTest extends UnitTestCase {
     $this->logger->expects($this->once())
       ->method('notice')->with("Purging Time's up time-sensitive cache-tags: timesup, timesup:weekly.");
 
-    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->state, $this->time, $this->loggerFactory);
+    $resolver = new WeeklyResolver($this->cacheTagsInvalidator, $this->configFactory, $this->state, $this->time, $this->loggerFactory);
     $resolver->purge();
   }